requirements.txt
absl-py==2.1.0
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.1.0
blessed==1.20.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cmake==3.31.4
confection==0.1.5
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
daam==0.2.0
diffusers==0.21.2
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
future==1.0.0
gpustat==1.1.1
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.23.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.5.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.7.0
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lightning-utilities==0.12.0
lit==18.1.8
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
MedPy==0.5.2
ml-collections==0.1.1
mmengine==0.10.4
mmsegmentation==1.2.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.2.1
nibabel==5.3.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open-clip-torch==1.2.1
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
preshed==3.0.9
prettytable==3.14.0
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pycocotools==2.0.8
pydantic==2.10.6
pydantic_core==2.27.2
pyDeprecate==0.3.1
pydicom==2.4.4
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==1.4.2
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.9.7
safetensors==0.5.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
SimpleITK==2.4.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.45.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.13.3
tomli==2.2.1
tomlkit==0.12.0
torch==1.11.0+cu113
torchaudio==0.11.0
torchmetrics==0.6.0
torchvision==0.12.0+cu113
tqdm==4.67.1
transformers==4.30.2
triton==3.2.0
typeguard==4.4.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
